requirements.txt
crewai[anthropic,google-genai]>=0.108.0
crewai-tools>=0.13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gradio>=5.22.0
